abstraction of hot vs. hot+warm storage

+++ C_Container/include/tracker_types.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <cstddef>
+#include <cstdint>
 #include <vector>
 
 namespace trk {
@@ -8,6 +9,6 @@ namespace trk {
 /**
  * @brief Common track ID list type used across pipeline stages and plugins.
  */
-using IdList = std::vector<std::size_t>;
+using IdList = std::vector<std::uint64_t>;
 
 } // namespace trk
